images workinginggit add -A

+++ server/db/models/alcohol.js
@@ -3,7 +3,7 @@ module.exports = (sequelize, DataTypes) => {
   const Alcohol = sequelize.define('Alcohol', {
     type: DataTypes.STRING,
     brandStyle: DataTypes.STRING,
-    sizeML: DataTypes.INTEGER,
+    sizeML: DataTypes.DECIMAL(10,0),
     sizeOZ: DataTypes.DECIMAL(10, 2),
     emptyBottleWeight: DataTypes.INTEGER,
     gramsPerOunce: DataTypes.DECIMAL(10, 2),
